Moves MdContent component to ui  and remove it when is  not used
