+++ pages/experiments/_experiment.vue
@@ -24,7 +24,7 @@ import { Context } from '@nuxt/types'
 import QiskitPage from '~/components/qiskit/QiskitPage.vue'
 import ExperimentHeader from '~/components/headers/ExperimentHeader.vue'
 import PageSection from '~/components/sections/PageSection.vue'
-import MdContent from '~/components/MdContent.vue'
+import MdContent from '~/components/ui/MdContent.vue'
 
 @Component({
   components: {
